# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Character AI API
  description: "\n# Character AI API\n\nA comprehensive API for managing interactive\
    \ AI characters in toys and devices.\n\n## Features\n\n- **Character Management**:\
    \ Create, configure, and manage AI characters\n- **Real-time Interaction**: Process\
    \ audio input and generate character responses\n- **Session Memory**: Maintain\
    \ conversation context with configurable limits\n- **Safety & Security**: Content\
    \ filtering and safety analysis\n- **Voice Management**: Handle voice artifacts\
    \ and embeddings\n- **Authentication**: Device identity and JWT-based authentication\n\
    \n## Authentication\n\nThe API supports multiple authentication methods:\n\n1.\
    \ **Device Registration**: Register a new device and receive authentication tokens\n\
    2. **JWT Tokens**: Use Bearer tokens for authenticated requests\n3. **Admin Tokens**:\
    \ Special tokens for administrative operations\n\n## Rate Limiting\n\n- Default:\
    \ 60 requests per minute per device\n- Burst allowance: 10 requests\n- Rate limit\
    \ headers are included in responses\n\n## Safety & Privacy\n\n- All content is\
    \ analyzed for toxicity and PII\n- Child-safe content filtering\n- Configurable\
    \ safety thresholds\n- Privacy-first design with on-device processing\n\n## Character\
    \ Profiles\n\nCharacters are defined using YAML profiles with:\n- Personality\
...
    \        "
  version: 1.0.0
servers:
...
paths:
...
  /api/v1/logs/search:
    post:
      tags:
      - log-search
      summary: Search Logs
      description: Search logs with advanced filtering and pagination.
      operationId: search_logs_api_v1_logs_search_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogSearchRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogSearchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/v1/logs/errors/summary:
    get:
      tags:
      - log-search
      summary: Get Error Summary
      description: Get error summary for the last N hours.
      operationId: get_error_summary_api_v1_logs_errors_summary_get
      security:
      - HTTPBearer: []
      parameters:
      - name: hours
        in: query
        required: false
        schema:
          type: integer
          maximum: 168
          minimum: 1
          default: 24
          title: Hours
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorSummaryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/logs/stats:
    get:
      tags:
      - log-search
      summary: Get Log Stats
      description: Get log aggregation statistics.
      operationId: get_log_stats_api_v1_logs_stats_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogStatsResponse'
      security:
      - HTTPBearer: []
...
  /api/v1/monitoring/grafana/export:
    post:
      tags:
      - monitoring
      summary: Export Grafana Configuration
      description: Export Grafana configuration files.
      operationId: export_grafana_configuration_api_v1_monitoring_grafana_export_post
      parameters:
      - name: output_dir
        in: query
        required: false
        schema:
          type: string
          default: monitoring/grafana
          title: Output Dir
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Export Grafana Configuration Api V1 Monitoring Grafana
                  Export Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/monitoring/elk/export:
    post:
      tags:
      - monitoring
      summary: Export Elk Configuration
      description: Export ELK stack configuration files.
      operationId: export_elk_configuration_api_v1_monitoring_elk_export_post
      parameters:
      - name: output_dir
        in: query
        required: false
        schema:
          type: string
          default: monitoring/elk
          title: Output Dir
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Export Elk Configuration Api V1 Monitoring Elk Export
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/monitoring/prometheus/export:
    post:
      tags:
      - monitoring
      summary: Export Prometheus Configuration
      description: Export Prometheus configuration files.
      operationId: export_prometheus_configuration_api_v1_monitoring_prometheus_export_post
      parameters:
      - name: output_dir
        in: query
        required: false
        schema:
          type: string
          default: monitoring/prometheus
          title: Output Dir
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Export Prometheus Configuration Api V1 Monitoring
                  Prometheus Export Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/monitoring/logs/search:
    get:
      tags:
      - monitoring
      summary: Search Logs
      description: Search logs using the log aggregation system.
      operationId: search_logs_api_v1_monitoring_logs_search_get
      parameters:
      - name: query
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Search query
          title: Query
        description: Search query
      - name: level
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/LogLevel'
          - type: 'null'
          description: Log level filter
          title: Level
        description: Log level filter
      - name: source
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/LogSource'
          - type: 'null'
          description: Log source filter
          title: Source
        description: Log source filter
      - name: start_time
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: Start time filter
          title: Start Time
        description: Start time filter
      - name: end_time
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: End time filter
          title: End Time
        description: End time filter
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          description: Maximum number of results
          default: 100
          title: Limit
        description: Maximum number of results
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Number of results to skip
          default: 0
          title: Offset
        description: Number of results to skip
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Logs Api V1 Monitoring Logs Search Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/monitoring/logs/correlation/{trace_id}:
    get:
      tags:
      - monitoring
      summary: Get Correlation Trace
      description: Get log correlation trace by trace ID.
      operationId: get_correlation_trace_api_v1_monitoring_logs_correlation__trace_id__get
      parameters:
      - name: trace_id
        in: path
        required: true
        schema:
          type: string
          title: Trace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Correlation Trace Api V1 Monitoring Logs Correlation  Trace
                  Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/monitoring/logs/errors/summary:
    get:
      tags:
      - monitoring
      summary: Get Error Summary
      description: Get error summary for the specified time period.
      operationId: get_error_summary_api_v1_monitoring_logs_errors_summary_get
      parameters:
      - name: hours
        in: query
        required: false
        schema:
          type: integer
          maximum: 168
          minimum: 1
          description: Hours to look back
          default: 24
          title: Hours
        description: Hours to look back
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Error Summary Api V1 Monitoring Logs Errors Summary
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/monitoring/export/all:
    post:
      tags:
      - monitoring
      summary: Export All Monitoring Configs
      description: Export all monitoring configurations (Grafana, ELK, Prometheus).
      operationId: export_all_monitoring_configs_api_v1_monitoring_export_all_post
      parameters:
      - name: output_dir
        in: query
        required: false
        schema:
          type: string
          default: monitoring/complete
          title: Output Dir
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Export All Monitoring Configs Api V1 Monitoring Export
                  All Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ErrorSummaryResponse:
      properties:
        total_errors:
          type: integer
          title: Total Errors
        by_component:
          additionalProperties:
            type: integer
          type: object
          title: By Component
        by_error_type:
          additionalProperties:
            type: integer
          type: object
          title: By Error Type
        by_device:
          additionalProperties:
            type: integer
          type: object
          title: By Device
        recent_errors:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Recent Errors
        time_range_hours:
          type: integer
          title: Time Range Hours
      type: object
      required:
      - total_errors
      - by_component
      - by_error_type
      - by_device
      - recent_errors
      - time_range_hours
      title: ErrorSummaryResponse
      description: Response model for error summary.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LogLevel:
      type: string
      enum:
      - DEBUG
      - INFO
      - WARNING
      - ERROR
      - CRITICAL
      title: LogLevel
      description: Log levels for filtering and analysis.
    LogSearchRequest:
      properties:
        start_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start Time
        end_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: End Time
        levels:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Levels
        sources:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Sources
        request_ids:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Request Ids
        trace_ids:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Trace Ids
        device_ids:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Device Ids
        character_ids:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Character Ids
        message_pattern:
          anyOf:
          - type: string
          - type: 'null'
          title: Message Pattern
        component_pattern:
          anyOf:
          - type: string
          - type: 'null'
          title: Component Pattern
        limit:
          type: integer
          maximum: 10000.0
          minimum: 1.0
          title: Limit
          default: 1000
        offset:
          type: integer
          minimum: 0.0
          title: Offset
          default: 0
        sort_by:
          type: string
          pattern: ^(timestamp|level|source)$
          title: Sort By
          default: timestamp
        sort_order:
          type: string
          pattern: ^(asc|desc)$
          title: Sort Order
          default: desc
      type: object
      title: LogSearchRequest
      description: Request model for log search.
    LogSearchResponse:
      properties:
        entries:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Entries
        total_count:
          type: integer
          title: Total Count
        execution_time_ms:
          type: number
          title: Execution Time Ms
        query:
          additionalProperties: true
          type: object
          title: Query
      type: object
      required:
      - entries
      - total_count
      - execution_time_ms
      - query
      title: LogSearchResponse
      description: Response model for log search.
    LogSource:
      type: string
      enum:
      - application
      - api
      - security
      - performance
      - safety
      - character
      - audio
      - llm
      - system
      title: LogSource
      description: Sources of log entries.
    LogStatsResponse:
      properties:
        stats:
          additionalProperties: true
          type: object
          title: Stats
        memory_logs_count:
          type: integer
          title: Memory Logs Count
        index_sizes:
          additionalProperties: true
          type: object
          title: Index Sizes
      type: object
      required:
      - stats
      - memory_logs_count
      - index_sizes
      title: LogStatsResponse
      description: Response model for log statistics.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from device registration
    AdminToken:
      type: apiKey
      in: header
      name: x-admin-token
      description: Admin token for administrative operations
...
security:
- BearerAuth: []
- AdminToken: []
...